requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
psycopg2
pandas
python-dotenv
